Fix INT_MAX sentinel collision in device tab sort
sortPagesByDeviceId() used INT_MAX as the sort key for tabs with no
valid device ID, but a legitimately valid device ID could itself be
INT_MAX, tying with the sentinel and letting std::stable_sort leave
the invalid tab before the valid one. Widen the sort key to long long
with LLONG_MAX as the sentinel so no valid int ID can collide with it.

Add a regression test covering an invalid-ID tab built before a
valid INT_MAX-ID tab.

+++ src/dialogs/adapterdevicesettings.cpp
@@ -122,13 +122,13 @@ AdapterDeviceSettings::AdapterDeviceSettings(SettingsModel* pSettingsModel, QWid
  */
 void AdapterDeviceSettings::sortPagesByDeviceId(QList<QWidget*>& pages, QStringList& names)
 {
-    QList<int> ids;
+    QList<long long> ids;
     ids.reserve(pages.size());
     for (auto* page : pages)
     {
         auto* tab = qobject_cast<DeviceConfigTab*>(page);
         const int id = tab ? tab->deviceId() : -1;
-        ids.append(id >= 0 ? id : INT_MAX);
+        ids.append(id >= 0 ? static_cast<long long>(id) : LLONG_MAX);
     }
 
     QList<int> order(pages.size());
